pyproject.toml
[project]
name = "markoun"
version = "0.1.0"
description = "A file-based Markdown editor"
authors = [{ name = "tropicalalgae", email = "tropicalalgae@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
dependencies = [
  "fastapi>=0.112.0,<0.113",
  "uvicorn==0.30.5",
  "pydantic[email]>=2.9.2,<3",
  "requests>=2.32.0,<3",
  "loguru>=0.7.0,<0.8",
  "pydantic-settings>=2.4.0,<3",
  "pytz~=2024.1",
  "pyyaml>=6.0.1,<7",
  "pymysql>=1.1.1,<2",
  "sqlalchemy>=2.0.32,<3",
  "passlib>=1.7.4,<2",
  "python-jose>=3.3.0,<4",
  "sqlmodel>=0.0.22,<0.0.23",
  "httpx>=0.28.1",
  "python-multipart>=0.0.20",
  "aiosqlite>=0.21.0",
  "asyncmy>=0.2.11",
  "cryptography>=46.0.3",
  "aiofiles>=25.1.0",
  "marko>=2.2.2",
]

[project.urls]
Repository = "https://github.com/tropical-algae/markoun.git"

[dependency-groups]
dev = [
  "pytest~=8.4.2",
  "pytest-cov>=7.0.0",
  "pytest-ordering>=0.6,<0.7",
  "black~=24.3",
  "autopep8>=2.0.0,<3",
  "ipdb>=0.13.0,<0.14",
  "mypy>=1.11.1,<2",
  "ruff>=0.5.6,<0.6",
  "types-pyyaml>=6.0.12.20240724,<7",
  "types-pytz>=2024.1.0.20240417,<2025",
  "sqlacodegen==3.0.0rc5",
  "pre-commit>=4.3.0,<5",
  "poethepoet>=0.37.0",
]
release = ["python-semantic-release>=10.3.1,<11"]
aws = ["mangum>=0.17.0,<0.18"]

[tool.uv]
default-groups = ["dev", "release", "aws"]

[[tool.uv.index]]
name = "proxy-source"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.hatch.build.targets.sdist]
include = ["src/markoun"]

[tool.hatch.build.targets.wheel]
include = ["src/markoun"]

[tool.hatch.build.targets.wheel.sources]
"src/markoun" = "markoun"

[tool.poe]
include = "poe_tasks.toml"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
